requirements.txt
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu117
# run above first!
numpy
pandas
scipy
tqdm
opencv-python
matplotlib
Pillow
jupyter
timm
ultralytics
scikit-image
visdom
imgaug
lmdb
imageio
openpyxl
pyyaml
fitter
segmentation-models-pytorch
imagecodecs
pycocotools
pyvips